requirements.txt
flask==1.0.2
flask-cors
geopy
datetime
Werkzeug
six
requests
tensorflow==2.3.0
albumentations==0.3.1
imutils==0.5.3
gunicorn
imageai
numpy
pandas
opencv-python-headless
unidecode
SQLAlchemy==1.3.8
PyMySQL==0.9.3
CMake==3.13.3
#dlib==19.19.0
face_recognition
face_recognition_models
Click>=6.0
Pillow
scipy>=0.17.0
google-cloud-storage
yagmail
exif
ffmpeg-python
ffprobe-python
ffprobe
torch==1.6.0+cu101
torchvision==0.7.0+cu101
google-cloud-vision
pygsheets
pyzbar